Add files via upload
In this project, the goal is to develop an intelligent system for answering TOEFL "Synonym Questions" by approximating the semantic similarity of word pairs. The system will calculate semantic descriptor vectors for given words using a provided text, where each coordinate represents the number of sentences in which the target word co-occurs with other words. The semantic descriptor vectors are stored as dictionaries, eliminating zeros for words that do not co-occur with the target word. The cosine similarity between these vectors serves as a measure of semantic similarity, enabling the system to identify the most closely related synonym from a list of alternatives.